# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Библиотека доступных API",
        "description": "Описание API для сервиса ЛЦТ 4 DeCODE-HS",
        "contact": {
            "url": "https://t.me/andrew_vaste"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://lct4.decodehs/api/",
            "description": "Получить список API"
        }
    ],
    "paths": {
...
        "excel/postamats/{id}/dashboard": {
            "post": {
                "tags": [
                    "Excel"
                ],
                "summary": "Выгрузить аналитику Постамата в xls",
                "description": "Выгрузить аналитику Постамата в xls",
                "operationId": "dashboard_postamat",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id постамата",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "excel/marketplaces/{id}/dashboard": {
            "post": {
                "tags": [
                    "Excel"
                ],
                "summary": "Выгрузить аналитику Маркетплейса в xls",
                "description": "Выгрузить аналитику Маркетплейса в xls",
                "operationId": "dashboard_marketplace",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id постамата",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
        "/reviews": {
            "get": {
                "tags": [
                    "Отзывы"
                ],
                "summary": "Получить все Отзывы",
                "description": "Получить все Отзывы",
                "operationId": "reviews",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Отзывы"
                ],
                "summary": "Создать Отзыв",
                "description": "Создать Отзыв",
                "operationId": "storeReview",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "text",
                                    "user_fio",
                                    "user_phone",
                                    "score"
                                ],
                                "properties": {
                                    "text": {
                                        "description": "Текст отзыва",
                                        "type": "string"
                                    },
                                    "user_fio": {
                                        "description": "ФИО автора",
                                        "type": "string"
                                    },
                                    "user_phone": {
                                        "description": "Телефон автора",
                                        "type": "string"
                                    },
                                    "score": {
                                        "description": "Оценка",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "postamat_id": {
                                        "description": "id Постамата",
                                        "type": "integer"
                                    },
                                    "thematic_id": {
                                        "description": "id Тематики",
                                        "type": "integer"
                                    },
                                    "source_id": {
                                        "description": "id Источника",
                                        "type": "integer"
                                    },
                                    "marketplace_id": {
                                        "description": "id Маркетплейса",
                                        "type": "integer"
                                    },
                                    "emotion_id": {
                                        "description": "id Эмоциональной окраски",
                                        "type": "integer"
                                    },
                                    "confirmed": {
                                        "description": "Отметка о проверке модерации",
                                        "type": "boolean"
                                    },
                                    "need_reaction": {
                                        "description": "Отметка о необходимости устранения",
                                        "type": "boolean"
                                    },
                                    "closed": {
                                        "description": "Отметка об устранении",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
    }
}